Changed to page router
