+++ pages/index.vue
@@ -1,0 +1,9 @@
+<template>
+  <div></div>
+</template>
+
+<script setup>
+const onClick = (e) => {
+  console.log("button works!");
+};
+</script>
